+++ shop/src/com/jinwei/service/ForderService.java
@@ -7,7 +7,7 @@ import com.jinwei.model.Forder;
 
 public interface ForderService extends BaseService<Forder> {
 	//计算购物车的总价格
-	public BigDecimal cluTotal(Forder forder);
+	public double cluTotal(Forder forder);
 	
 	public void updateStatusById(int id, int sid);
 }
